pyproject.toml
[tool.poetry]
name = "notebook-repo"
version = "0.1.0"
description = ""
authors = ["guilherme <guilherme-lopes58@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
respx = "^0.21.1"
time-machine = "^2.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
